+++ java/org_libjpegturbo_turbojpeg_TJDecompressor.h
@@ -114,10 +114,10 @@ JNIEXPORT void JNICALL Java_org_libjpegturbo_turbojpeg_TJDecompressor_decodeYUV8
 /*
  * Class:     org_libjpegturbo_turbojpeg_TJDecompressor
  * Method:    saveImage
- * Signature: (ILjava/lang/String;Ljava/lang/Object;IIII)V
+ * Signature: (ILjava/lang/String;Ljava/lang/Object;IIIIII)V
  */
 JNIEXPORT void JNICALL Java_org_libjpegturbo_turbojpeg_TJDecompressor_saveImage
-  (JNIEnv *, jobject, jint, jstring, jobject, jint, jint, jint, jint);
+  (JNIEnv *, jobject, jint, jstring, jobject, jint, jint, jint, jint, jint, jint);
 
 #ifdef __cplusplus
 }