+++ ChangeLog.md
@@ -51,6 +51,13 @@ future release.  TJBench command-line arguments can now be abbreviated as well.
 (Where possible, the abbreviations are the same as those supported by cjpeg,
 djpeg, and jpegtran.)
 
+7. The TurboJPEG C and Java APIs have been improved in the following ways:
+
+     - New image I/O methods (`TJCompressor.loadSourceImage()` and
+`TJDecompressor.saveImage()`) have been added to the Java API.  These methods
+work similarly to the `tj3LoadImage*()` and `tj3SaveImage*()` functions in the
+C API.
+
 
 3.0.4
 =====